Extract CLI argument-group builders
